Improve: singular points APIs
