+++ src/routes.py
@@ -3,10 +3,9 @@ from flask_restful import Api
 
 from src.bacnet_server.resources.mapping.mapping import BPGPMappingResourceList, GBPMappingResourceByGenericPointUUID, \
     GBPMappingResourceByBACnetPointUUID, BPGPMappingResourceByUUID
-from src.bacnet_server.resources.point.point_name import BACnetPointName
-from src.bacnet_server.resources.point.point_object import BACnetPointObject
 from src.bacnet_server.resources.point.point_plural import BACnetPointPlural
-from src.bacnet_server.resources.point.point_singular import BACnetPointSingular
+from src.bacnet_server.resources.point.point_singular import BACnetPointSingularByUUID, \
+    BACnetPointSingularByName, BACnetPointSingularByObject
 from src.bacnet_server.resources.point.point_sync import BPGPSync
 from src.bacnet_server.resources.server.server import BACnetServer
 from src.system.resources.ping import Ping
@@ -16,9 +15,9 @@ api_bacnet_server = Api(bp_bacnet_server)
 
 api_bacnet_server.add_resource(BACnetServer, '/server')
 api_bacnet_server.add_resource(BACnetPointPlural, '/points')
-api_bacnet_server.add_resource(BACnetPointSingular, '/points/uuid/<string:uuid>')
-api_bacnet_server.add_resource(BACnetPointObject, '/points/obj/<string:object_type>/<string:address>')
-api_bacnet_server.add_resource(BACnetPointName, '/points/name/<string:object_name>')
+api_bacnet_server.add_resource(BACnetPointSingularByUUID, '/points/uuid/<string:uuid>')
+api_bacnet_server.add_resource(BACnetPointSingularByObject, '/points/obj/<string:object_type>/<string:address>')
+api_bacnet_server.add_resource(BACnetPointSingularByName, '/points/name/<string:object_name>')
 
 # BACnet points <> Generic points mappings
 bp_mapping_bp_gp = Blueprint('mappings_bp_gp', __name__, url_prefix='/api/mappings/bp_gp')
